requirements.txt
amqp==2.5.1
arrow==0.14.5
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.2.0
Babel==2.7.0
billiard==3.6.1.0
boto==2.49.0
boto3==1.9.212
botocore==1.12.246
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
codacy-coverage==1.3.11
coverage==4.5.4
cssselect==1.1.0
Django==2.2.4
django-appconf==1.0.3
django-compressor==2.3
django-cors-headers==3.2.1
django-filter==2.2.0
django-haystack==2.8.1
django-phonenumber-field==2.2.0
django-simple-pagination==1.3
django-smtp-ssl==1.0
django-storages==1.7
djangorestframework==3.10.3
docutils==0.15.2
elasticsearch==5.5.3
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.8
importlib-metadata==0.23
isort==4.3.21
jdcal==1.4.1
jmespath==0.9.4
joblib==0.14.1
kombu==4.6.5
lazy-object-proxy==1.4.3
lxml==4.4.1
Markdown==3.1.1
mccabe==0.6.1
more-itertools==7.2.0
numpy==1.18.0
openpyxl==2.6.3
packaging==19.2
pandas==0.25.3
pdfkit==0.6.1
phonenumbers==8.10.4
pluggy==0.13.0
psycopg2-binary==2.8.3
py==1.8.0
pylint==2.4.4
pyparsing==2.4.2
pytest==5.1.1
pytest-django==3.5.1
python-dateutil==2.8.0
python-memcached==1.59
pytz==2019.3
raven==6.10.0
rcssmin==1.0.6
redis==3.3.8
requests==2.22.0
rjsmin==1.1.0
s3transfer==0.2.1
scikit-learn==0.22.1
scipy==1.4.1
six==1.12.0
sklearn==0.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.6
uWSGI==2.0.18
vine==1.3.0
wcwidth==0.1.7
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
